chore: move configuration.ts to libs
